# Improve @turf/boolean-overlap performance on MultiPoints